More timeline fixes
Fixes issues where posts were loading in timelines that are selected, for instance when the notifications timeline was selected, if a new post came in, it would show up as a normal post in that timeline. This should no longer happen.
Timelines should now load more smoothly, and switching between them is seemless.